test(cockpit): make the five worker mocks real, close P2 test-integrity item
Formatter, diff, xml, typescript, and refactoring worker tests were passing
against a no-op mock that discarded every postMessage() and never resolved —
not one test in those five tool suites asserted on real worker output.
Extracts each worker's pure logic into a sibling *.api.ts module (mirroring
the existing regex.api.ts split) so both the real worker and a new live mock
run identical code. Adds round-trip tests that fail if any mock reverts to a
no-op, plus a dynamic-import alias-shadowing guard, and corrects the earlier
P0 item's overstated round-trip coverage claim.
