+++ apps/cockpit/src/tools/__tests__/refactoring-toolkit.test.tsx
@@ -1,5 +1,5 @@
 import { describe, expect, it } from 'vitest'
-import { screen } from '@testing-library/react'
+import { fireEvent, screen, waitFor } from '@testing-library/react'
 import { renderTool } from './test-utils'
 import RefactoringToolkit from '../refactoring-toolkit/RefactoringToolkit'
 
@@ -26,5 +26,38 @@ describe('RefactoringToolkit', () => {
   it('renders language selector', () => {
     renderTool(RefactoringToolkit)
     expect(screen.getByDisplayValue('JavaScript')).toBeInTheDocument()
+  })
+
+  // ── Worker round-trip ────────────────────────────────────────────
+  // Preview generation is auto-triggered (debounced 300ms) via the real
+  // jscodeshift transform in the worker mock. A no-op worker mock never
+  // resolves applyTransforms() and the Apply button / diff preview never
+  // appear.
+
+  it('applies the var-to-const transform via the real refactoring worker', async () => {
+    renderTool(RefactoringToolkit)
+    const editor = screen.getByTestId('monaco-editor')
+
+    fireEvent.change(editor, { target: { value: 'var x = 1;' } })
+
+    const transformLabel = screen.getByText('var → const/let')
+    const checkbox = transformLabel.closest('label')?.querySelector('input[type="checkbox"]')
+    expect(checkbox).toBeTruthy()
+    fireEvent.click(checkbox as Element)
+
+    const applyButton = await waitFor(
+      () => {
+        const button = screen.getByText('Apply')
+        return button
+      },
+      { timeout: 3000 }
+    )
+    expect(screen.getByTestId('modified-editor')).toHaveValue('const x = 1;')
+
+    fireEvent.click(applyButton)
+
+    await waitFor(() => {
+      expect(screen.getByTestId('monaco-editor')).toHaveValue('const x = 1;')
+    })
   })
 })
