+++ apps/cockpit/src/workers/diff.worker.ts
@@ -1,28 +1,10 @@
 import { handleRpc } from './rpc'
-import { createTwoFilesPatch } from 'diff'
-
-type DiffOptions = {
-  ignoreWhitespace?: boolean
-  jsonMode?: boolean
-}
+import { computeDiff } from './diff.api'
+import type { DiffOptions } from './diff.api'
 
 const api = {
   computeDiff(left: string, right: string, options: DiffOptions = {}): string {
-    let a = left
-    let b = right
-
-    if (options.jsonMode) {
-      try {
-        a = JSON.stringify(JSON.parse(a), null, 2)
-        b = JSON.stringify(JSON.parse(b), null, 2)
-      } catch {
-        // If not valid JSON, diff as-is
-      }
-    }
-
-    return createTwoFilesPatch('left', 'right', a, b, undefined, undefined, {
-      ignoreWhitespace: options.ignoreWhitespace,
-    })
+    return computeDiff(left, right, options)
   },
 }
